+++ modules/cda/controller/CDAPatient.ts
@@ -35,7 +35,7 @@ import {
     CDA as CDAConfig
 } from '../../../config.private';
 import {
-    ConfiguracionPrestacionModel
+    configuracionPrestacionModel
 } from './../../../core/term/schemas/configuracionPrestacion';
 
 /**
@@ -124,7 +124,7 @@ let rootOID = CDAConfig.rootOID;
 export async function matchCode(snomed) {
     let prestacion: any;
     if (!isNaN(snomed)) {
-        prestacion = await ConfiguracionPrestacionModel.findOne({
+        prestacion = await configuracionPrestacionModel.findOne({
             'snomed.conceptId': snomed
         });
         if (prestacion) {
@@ -143,7 +143,7 @@ export async function matchCode(snomed) {
  */
 
 export async function matchCodeByLoinc(loinc) {
-    let prestacion: any = await ConfiguracionPrestacionModel.find({
+    let prestacion: any = await configuracionPrestacionModel.find({
         'loinc.code': loinc
     });
     if (prestacion.length > 0) {
