+++ modules/turnos/controller/operationsCacheController/operationsAgenda.ts
@@ -4,7 +4,7 @@ import {
     agendasCache
 } from '../../../legacy/schemas/agendasCache';
 import { tipoPrestacion } from './../../../../core/tm/schemas/tipoPrestacion';
-import { configuracionPrestacionModel } from './../../../../core/term/schemas/configuracionPrestaciones';
+import { configuracionPrestacionModel } from './../../../../core/term/schemas/configuracionPrestacion';
 import * as sql from 'mssql';
 import * as moment from 'moment';
 import * as pacientes from './../../../../core/mpi/controller/paciente';
